Do not copy the 3D stack in memory if the source image is not virtual.
For large movies it is a good idea to copy a full 3D+C stack on
a new Img. Why? Because when movies are very big they will be
opened as a virtual stack in ImageJ. Virtual stacks are cool,
but a disk access happens every-time we change the Z position.
And when we do the extract of the surface, we iterate along Z
for each XY position. This copy here works like a simple cache
for 3D.

But it takes time, which might slow down process when we have
only one time-point, but a large 3D image that fits into memory.
So we let the user specifies whether the source movie is virtual
(do a copy) or not (no copy needed, the default).