Update display (min  & max range and time) during computation.
So that the user is not presented with black images.
Now we have a composite image, with min & max taken from the
input image for  the projection.
For the reference image, we take the min & max from the max number
of z-slices.
There is still a bug where the displais is not fully refreshed on
the first time-point, but this one I do not understand.