Release v1.2.0.
- Fix bug: The source image was modified (smoothed) if the binning
was left to 1 and the gaussian prefilter value different from 0.

- Added two new methods to detect reference surface:
	sparse max of mean and
	sparse max of std.
They work the same as their non-sparse counterpart. However, we
do not compute the filter value for every pixel of the source slice,
but only a grid spaced by the size of the filter.
And we multithread it. Greatly reduce omputation time for large
images for which binning is not adequate.

- Speed-up projection step for images that are stored in memory
(as opposed to virtual stack that are streamed slice by slice
from disk).

- Update display (min  & max range and time) during computation.
So that the user is not presented with black images.
Now we have a composite image, with min & max taken from the
input image for  the projection.
For the reference image, we take the min & max from the max number
of z-slices.
There is still a bug where the displais is not fully refreshed on
the first time-point, but this one I do not understand.

- Fix current time-point not properly retrieved for preview.

- Implement parameter persistence.
The parameter values are remembered between Fiji sessions.