Use a new one-pass approach for the projection.
This one makes processing of images larger than RAM possible,
simply using the Fiji virtual stack thingie. It is long, but possible.

On 'normal' images, it is about 15% slower: For a
1148 x 1148 x 26 x 2 images, it takes about 6.30s for the whole process,
against 5.67s before. But this is worth it.

Also, we cannot do median projection any more, but again, this is
worth it.